Move configuration enums to shared location
- Create ConfigurationEnums.swift with AuthenticationMode and TitleMode
- Remove duplicate enum definitions from UI files
- Fix ConfigManager to use correct enum case names (.osAuth instead of .os)
- Add description property to AuthenticationMode for UI display
- Proper separation of shared types from UI-specific code
